+++ nix/hydra/demo.nix
@@ -211,6 +211,36 @@
               # Ensure test starts only after devnet has been prepared
               depends_on."prepare-devnet".condition = "process_completed";
             };
+            node-exporter = {
+              command = "${pkgs.prometheus-node-exporter}/bin/node_exporter --web.listen-address=:9100";
+            };
+            prometheus = {
+              working_dir = "./demo";
+              command = ''
+                ${pkgs.prometheus}/bin/prometheus \
+                    --config.file=./prometheus/nix/prometheus.yml \
+                    --storage.tsdb.path=./devnet/prometheus \
+                    --web.listen-address=127.0.0.1:9090
+              '';
+              depends_on = {
+                "node-exporter".condition = "process_started";
+                "hydra-node-alice".condition = "process_started";
+                "hydra-node-bob".condition = "process_started";
+                "hydra-node-carol".condition = "process_started";
+              };
+            };
+            grafana = {
+              working_dir = "./demo";
+              command = ''
+                ${pkgs.grafana}/bin/grafana server \
+                    --homepath ${pkgs.grafana}/share/grafana \
+                    cfg:default.paths.data=$(pwd)/devnet/grafana/data \
+                    cfg:default.paths.logs=$(pwd)/devnet/grafana/logs \
+                    cfg:default.paths.plugins=$(pwd)/devnet/grafana/plugins \
+                    cfg:default.paths.provisioning=$(pwd)/grafana/nix
+              '';
+              depends_on."prometheus".condition = "process_started";
+            };
           };
         };
       };